03_quant_trading/04_stat_arb_trader_execution/requirements.txt
# Elite Statistical Arbitrage - Minimal Dependencies

# Core
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0
scipy>=1.10.0,<2.0.0

# Statistical models
statsmodels>=0.13.0,<0.15.0
scikit-learn>=1.3.0,<2.0.0
hmmlearn>=0.3.0,<0.4.0

# Data
yfinance>=0.2.28

# Visualization
matplotlib>=3.7.0,<4.0.0

# Config
pyyaml>=6.0
